CRAS: Make shared structures architecture-independent.
This allows x86 clients to connect to x86-64 server: mainly useful
for crouton as users want to play sound from x86 binaries running
in their x86-64 chroots.

Get rid of size_t, long, and timespec, whose size depend on the
architecture. We also pack all structures to make sure alignment
is the same across platforms.

We assume that sizeof(enum) = sizeof(int) = 4, which is true on
x86/x86-64/arm with GCC (using default compilation flags).

Volume is now an uint32_t (was size_t), capture gain is now an
int32_t (was long): no difference on x86/arm, and 32-bit is more than
enough range on x86-64. Frame rate, number of channels, and shm
offsets are also converted from size_t to uint32_t.

We replace timespec by a fixed-size cras_timespec, where both fields
are 32-bit signed integers (as on x86/arm, x86-64 uses 64-bit ints).
All we care about in CRAS is differences between 2 cras_timespec, so
64-bit fields are not required (even after year 2038...).

BUG=chromium:309961
TEST=Audio input/output still works (e.g. try a Hangout session)
TEST=In crosh, developer mode: 'cd; arecord -f dat -d 5 > x; aplay x'
     records 5 seconds of audio, then plays it back.
CQ-DEPEND=CL:189206

+++ cras/src/tests/alsa_io_unittest.cc
@@ -986,13 +986,13 @@ const char *cras_alsa_mixer_get_output_name(
 }
 
 //  From system_state.
-size_t cras_system_get_volume()
+uint32_t cras_system_get_volume()
 {
   sys_get_volume_called++;
   return sys_get_volume_return_value;
 }
 
-long cras_system_get_capture_gain()
+int32_t cras_system_get_capture_gain()
 {
   sys_get_capture_gain_called++;
   return sys_get_capture_gain_return_value;
@@ -1010,12 +1010,12 @@ int cras_system_get_capture_mute()
   return sys_get_capture_mute_return_value;
 }
 
-void cras_system_set_volume_limits(long min, long max)
+void cras_system_set_volume_limits(int32_t min, int32_t max)
 {
   sys_set_volume_limits_called++;
 }
 
-void cras_system_set_capture_gain_limits(long min, long max)
+void cras_system_set_capture_gain_limits(int32_t min, int32_t max)
 {
   sys_set_capture_gain_limits_called++;
 }
